requirements.txt
Flask
google-genai
python-dotenv
twilio
gunicorn
